+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 
 - Connection.CloseGraceful() unlike Connection.Close() waits for all
   requests to complete (#257)
+- ConnectionPool.CloseGraceful() unlike ConnectionPool.Close() waits for all
+  requests to complete (#257)
 
 ### Changed